+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -216,7 +216,6 @@ module ActiveRecord
         float:       { name: "float" },
         decimal:     { name: "decimal" },
         datetime:    { name: "timestamp" },
-        timestamp:   { name: "timestamp" },
         time:        { name: "time" },
         date:        { name: "date" },
         daterange:   { name: "daterange" },