+++ activerecord/lib/active_record/connection_adapters/sqlite3_adapter.rb
@@ -69,7 +69,6 @@ module ActiveRecord
         float:        { name: "float" },
         decimal:      { name: "decimal" },
         datetime:     { name: "datetime" },
-        timestamp:    { name: "datetime" },
         time:         { name: "time" },
         date:         { name: "date" },
         binary:       { name: "blob" },